Move CSSReporter from script_layout_interface to script
