+++ components/script_layout_interface/Cargo.toml
@@ -21,7 +21,6 @@ gfx_traits = {path = "../gfx_traits"}
 html5ever = "0.22"
 ipc-channel = "0.11"
 libc = "0.2"
-log = "0.4"
 time = "0.1.17"
 malloc_size_of = { path = "../malloc_size_of" }
 malloc_size_of_derive = "0.1"
